wip: consolidate into route
